% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_mem_hash.R
\name{attach_mem_hash}
\alias{attach_mem_hash}
\title{join information of the memoised output with the input in the cache}
\usage{
attach_mem_hash(x, db, f, ...)
}
\arguments{
\item{x}{memoised output}

\item{db}{cache object}

\item{f}{memoised function}

\item{...}{parameters passed to f}
}
\value{
nothing
}
\description{
join information of the memoised output with the input in the cache
}
\details{
rds file written to db$path with the same name as the memoised output,
but with a leading `_`.
}
\author{
Jonathan Sidi
}
