% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digest_dictionary.R
\name{digest_dictionary}
\alias{digest_dictionary}
\title{create/update a remote digest dictionary}
\usage{
digest_dictionary(..., storage = "storage")
}
\arguments{
\item{...}{objects to digest and add to dictionary}
}
\value{
character, named vector of the saved digests and the names of the objects they represent
}
\description{
maintain a dictionary of a digest on a remote repository
}
\examples{
digest_dictionary(mtcars)
digest_dictionary(iris)
digest_dictionary(iris,iris3)

rm(list=ls(pattern = '^_'))

digest_dictionary(mtcars,iris,iris3)

}
\author{
Jonathan Sidi
}
