% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cache.R
\name{get_cache}
\alias{get_cache}
\title{return the digest associated with a memoised call}
\usage{
get_cache(f, ...)
}
\arguments{
\item{f}{memoised function}

\item{...}{parameters passed to f}
}
\value{
character
}
\description{
return the digest associated with a memoised call
}
\author{
Jonathan Sidi
}
