% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_mem.R
\name{init_mem}
\alias{init_mem}
\title{wrapper to initialize memoise}
\usage{
init_mem(f, cache = cache_memory(), fname = "mem_f")
}
\arguments{
\item{f}{function, Function of which to create a memoised copy}

\item{cache}{function, Cache function, Default: cache_memory()}

\item{fname}{character, name of function to assign in environment, Default: 'mem_f'}
}
\value{
memoised function
}
\description{
wrapper to initialize memoise in parent environment
}
\seealso{
\code{\link[digest]{digest}}
 \code{\link[memoise]{memoise}}
}
\author{
Jonathan Sidi
}
