% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dictionary.R
\name{load_dictionary}
\alias{load_dictionary}
\title{Load the digest dictionary to the global environment}
\usage{
load_dictionary(storage = "storage")
}
\arguments{
\item{storage}{character, path to dictionary file, Default: 'storage'}
}
\description{
Load the digest dictionary to the global environment
}
\examples{
load_dictionary()
dictionary
}
\author{
Jonathan Sidi
}
