% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack_cache.R
\name{unpack_cache}
\alias{unpack_cache}
\title{unpack a memoised function}
\usage{
unpack_cache(f, ...)
}
\arguments{
\item{f}{memoised function}

\item{...}{parameters passed to f}
}
\value{
list of internal objects associated with the memoised function
}
\description{
unpack elements from the body of the main memoise function call
}
\author{
Jonathan Sidi
}
