% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dictionary.R
\name{load_dictionary}
\alias{load_dictionary}
\title{Load the digest dictionary to the global environment}
\usage{
load_dictionary(local_storage = "storage", remote_storage = "storage")
}
\arguments{
\item{local_storage}{character, subdirectory where digests are saved locally, Default: 'storage'}

\item{remote_storage}{character, subdirectory where digests are saved remotely, Default: 'storage'}
}
\description{
Load the digest dictionary to the global environment
}
\examples{
load_dictionary()
dictionary
}
\author{
Jonathan Sidi
}
