% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memofit.R
\name{memofit}
\alias{memofit}
\title{FUNCTION_TITLE}
\usage{
memofit(data, db, f)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{db}{PARAM_DESCRIPTION}

\item{f}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[hbgd]{get_avail_methods}}
}
\author{
Jonathan Sidi
}
