% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synch_remote.R
\name{synch_remote}
\alias{synch_remote}
\title{Retrieve current files in remote storage to local directory}
\usage{
synch_remote(repo, local_storage = "storage", remote_storage = "storage",
  ...)
}
\arguments{
\item{repo}{character, user/repository}

\item{local_storage}{character, subdirectory where digests are saved locally, Default: 'storage'}

\item{remote_storage}{character, subdirectory where digests are saved remotely, Default: 'storage'}

\item{...}{parameters to pass to \code{\link[vcs]{ls_remote}}}
}
\value{
nothing
}
\description{
FUNCTION_DESCRIPTION
}
\seealso{
\code{\link[vcs]{ls_remote}}
}
\author{
Jonathan Sidi
}
